requirements.txt
# Core dependencies
streamlit>=1.28.0
azure-ai-projects>=1.0.0b11
azure-identity>=1.15.0
azure-cosmos>=4.5.1
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.1.0
reportlab>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
altair>=5.0.0
pydeck>=0.8.0

# Enhanced reliability and performance
redis>=4.0.0
celery>=5.2.0
websockets>=10.0
prometheus-client>=0.14.0

# Testing and quality assurance
pytest>=7.0.0
pytest-asyncio>=0.19.0
locust>=2.10.0

# Machine learning (for future analytics)
scikit-learn>=1.1.0

# Monitoring and observability
jaeger-client>=4.8.0
